modify the warnings

+++ src/main/java/io/xdag/mine/miner/MinerCalculate.java
@@ -99,8 +99,7 @@ public class MinerCalculate {
         .append("   unpaid: ")
         .append(String.format("%.6f", unpaid))
         .append("   HashRate: ")
-        .append(
-            String.format("%.6f", BasicUtils.xdag_log_difficulty2hashrate(miner.getMeanLogDiff())))
+        .append(String.format("%.6f", BasicUtils.xdag_log_difficulty2hashrate(miner.getMeanLogDiff())))
         .append("\n");
 
     Map<InetSocketAddress, MinerChannel> channels = miner.getChannels();
@@ -234,7 +233,7 @@ public class MinerCalculate {
 
   public static double welfordOnePass(double mean, double sample, int nsamples) {
     if (nsamples > 0) {
-      double temp = BigDecimalUtils.div(BigDecimalUtils.sub(sample, mean), (double) nsamples);
+      double temp = BigDecimalUtils.div(BigDecimalUtils.sub(sample, mean),  nsamples);
       mean = BigDecimalUtils.add(mean, temp);
     }
